arch/arm/src/samd2l2/sam_usb.c: Fix type mismatches

+++ arch/arm/src/samd2l2/sam_usb.c
@@ -783,7 +783,7 @@ static void sam_putreg32(uint32_t regval, uintptr_t regaddr)
   putreg32(regval, regaddr);
 }
 #else
-static inline void sam_putreg32(uint32_t regval, uint32_t regaddr)
+static inline void sam_putreg32(uint32_t regval, uintptr_t regaddr)
 {
   putreg32(regval, regaddr);
 }
@@ -836,7 +836,7 @@ static void sam_putreg16(uint16_t regval, uintptr_t regaddr)
   putreg16(regval, regaddr);
 }
 #else
-static inline void sam_putreg16(uint16_t regval, uint32_t regaddr)
+static inline void sam_putreg16(uint16_t regval, uintptr_t regaddr)
 {
   putreg16(regval, regaddr);
 }
@@ -889,7 +889,7 @@ static void sam_putreg8(uint8_t regval, uintptr_t regaddr)
   putreg8(regval, regaddr);
 }
 #else
-static inline void sam_putreg8(uint8_t regval, uint32_t regaddr)
+static inline void sam_putreg8(uint8_t regval, uintptr_t regaddr)
 {
   putreg8(regval, regaddr);
 }
